minor dtype refactor
