+++ crates/nu_plugin_polars/src/dataframe/values/nu_dtype/mod.rs
@@ -2,12 +2,12 @@ pub mod custom_value;
 
 use custom_value::NuDataTypeCustomValue;
 use nu_protocol::{record, ShellError, Span, Value};
-use polars::prelude::{DataType, PlSmallStr, TimeUnit, UnknownKind};
+use polars::prelude::{DataType, Field, PlSmallStr, TimeUnit, UnknownKind};
 use uuid::Uuid;
 
 use crate::{Cacheable, PolarsPlugin};
 
-use super::{nu_schema::dtype_to_value, CustomValueSupport, PolarsPluginObject, PolarsPluginType};
+use super::{CustomValueSupport, PolarsPluginObject, PolarsPluginType};
 
 #[derive(Debug, Clone)]
 pub struct NuDataType {
@@ -300,6 +300,18 @@ pub fn str_to_dtype(dtype: &str, span: Span) -> Result<DataType, ShellError> {
     }
 }
 
+pub(crate) fn fields_to_value(fields: impl Iterator<Item = Field>, span: Span) -> Value {
+    let record = fields
+        .map(|field| {
+            let col = field.name().to_string();
+            let val = dtype_to_value(field.dtype(), span);
+            (col, val)
+        })
+        .collect();
+
+    Value::record(record, Span::unknown())
+}
+
 fn str_to_time_unit(ts_string: &str, span: Span) -> Result<TimeUnit, ShellError> {
     match ts_string {
         "ms" => Ok(TimeUnit::Milliseconds),
@@ -312,5 +324,12 @@ fn str_to_time_unit(ts_string: &str, span: Span) -> Result<TimeUnit, ShellError>
             help: None,
             inner: vec![],
         }),
+    }
+}
+
+pub(crate) fn dtype_to_value(dtype: &DataType, span: Span) -> Value {
+    match dtype {
+        DataType::Struct(fields) => fields_to_value(fields.iter().cloned(), span),
+        _ => Value::string(dtype.to_string().replace('[', "<").replace(']', ">"), span),
     }
 }
